# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "재고관리 + 주문처리 + 스테이블코인 정산 API",
        "title": "B2B Commerce Settlement Engine API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/users/{id}/wallets": {
            "get": {
                "description": "Get all wallets for a user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "List user wallets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User external ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wallet list",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_wallet.ListWalletsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid UUID format",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Register a new Ethereum wallet for the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Register a new wallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User external ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Wallet registration data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_wallet.RegisterWalletRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Wallet created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_wallet.WalletResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Wallet address already registered",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}/wallets/{walletId}": {
            "get": {
                "description": "Retrieve wallet details by external ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Get wallet by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User external ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Wallet external ID (UUID)",
                        "name": "walletId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wallet details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_wallet.WalletResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid UUID format",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Wallet not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a non-primary wallet (hard delete)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Delete wallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User external ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Wallet external ID (UUID)",
                        "name": "walletId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Wallet deleted"
                    },
                    "400": {
                        "description": "Cannot delete primary wallet",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Wallet not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}/wallets/{walletId}/label": {
            "put": {
                "description": "Update the label of a wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Update wallet label",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User external ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Wallet external ID (UUID)",
                        "name": "walletId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Label update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_wallet.UpdateLabelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated wallet",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_wallet.WalletResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Wallet not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}/wallets/{walletId}/set-primary": {
            "post": {
                "description": "Set a verified wallet as the primary wallet",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Set wallet as primary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User external ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Wallet external ID (UUID)",
                        "name": "walletId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Primary wallet",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_wallet.WalletResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Wallet not verified",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Wallet not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}/wallets/{walletId}/verify": {
            "post": {
                "description": "Verify wallet ownership using EIP-712 signature",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Verify wallet ownership",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User external ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Wallet external ID (UUID)",
                        "name": "walletId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Signature and message data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_wallet.VerifyWalletRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verified wallet",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_wallet.WalletResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid signature or verification failed",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Wallet not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.ErrorBody"
                }
            }
        },
        "github_com_ahwlsqja_StableCoin-B2B-Commerce-Settlement-Engine_internal_common_middleware.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
...
        "internal_wallet.ListWalletsResponse": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                },
                "wallets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_wallet.WalletResponse"
                    }
                }
            }
        },
        "internal_wallet.RegisterWalletRequest": {
            "type": "object",
            "required": [
                "address"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "example": "0x742d35Cc6634C0532925a3b844Bc454e4438f44e"
                },
                "label": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "My Main Wallet"
                }
            }
        },
        "internal_wallet.UpdateLabelRequest": {
            "type": "object",
            "required": [
                "label"
            ],
            "properties": {
                "label": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "Trading Wallet"
                }
            }
        },
        "internal_wallet.VerifyWalletRequest": {
            "type": "object",
            "required": [
                "message",
                "signature"
            ],
            "properties": {
                "message": {
                    "$ref": "#/definitions/internal_wallet.VerifyWalletRequestMessage"
                },
                "signature": {
                    "description": "Signature: 0x prefix + 130 hex chars (65 bytes)",
                    "type": "string",
                    "example": "0x1234...abcd"
                }
            }
        },
        "internal_wallet.VerifyWalletRequestMessage": {
            "type": "object",
            "required": [
                "nonce",
                "timestamp"
            ],
            "properties": {
                "nonce": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 8,
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "timestamp": {
                    "type": "integer",
                    "example": 1706000000
                }
            }
        },
        "internal_wallet.WalletResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "0x742d35cc6634c0532925a3b844bc454e4438f44e"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "is_primary": {
                    "type": "boolean",
                    "example": false
                },
                "is_verified": {
                    "type": "boolean",
                    "example": false
                },
                "label": {
                    "type": "string",
                    "example": "My Main Wallet"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}